requirements.txt
alembic==1.14.0
aniso8601==10.0.0
apispec==6.8.1
APScheduler==3.11.0
attrs==24.3.0
Authlib==1.4.0
bcrypt==4.2.1
bidict==0.23.1
blinker==1.9.0
cachelib==0.13.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
flasgger==0.9.7.1
Flask==2.0.2
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-Dance==7.0.0
Flask-JWT-Extended==4.3.1
Flask-Migrate==3.1.0
Flask-RESTful==0.3.10
flask-restx==1.3.0
flask-smorest==0.45.0
Flask-SocketIO==5.5.0
Flask-SQLAlchemy==2.5.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Mako==1.3.8
MarkupSafe==3.0.2
marshmallow==3.25.0
mistune==3.1.0
mysqlclient==2.2.1
oauthlib==2.1.0 
packaging==24.2
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-engineio==4.11.1
python-socketio==5.12.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==1.1.0
rpds-py==0.22.3
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==1.4.44
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.3.0
URLObject==2.4.3
webargs==8.6.0
Werkzeug==2.0.3
wsproto==1.2.0
